Fix Compile Update Mastery Stats -[Core/Spell] Mage - Fixed Permafrost -[Core/Spell] Warrior - Fixed Lambs of the Slaughter -[Core/Mmaps] Fixed Creature swimming MMaps Issue

+++ src/server/game/DataStores/DBCStructure.h
@@ -1698,8 +1698,7 @@ struct SpellEffectEntry
     uint32    EffectImplicitTargetB;                        // 23        m_implicitTargetB
     uint32    EffectSpellId;                                // 24        new 4.0.0
     uint32    EffectIndex;                                  // 25        new 4.0.0
-    uint32    EffectBonusCoefficient;                       // 26        new 4.0.0                        
-    //uint32  Unk0                                          // 27        4.2.0 only 0 or 1
+    float     EffectBonusCoefficient;                        // 26        new 4.0.0                        
 };
 
 #define MAX_SPELL_EFFECTS 3
